+++ Changelog.md
@@ -12,3 +12,4 @@
 - Support conversion for annotation constructor calls
 - Place comments from the middle of the call to the end
 - Drop line breaks between operator arguments (except '+', "-", "&&" and "||")
+- Add non-null assertions on call site for non-null parameters